+++ wattio/switch.py
@@ -13,12 +13,14 @@ from homeassistant.util.json import load_json, save_json
 from homeassistant.helpers.entity import Entity
 from homeassistant.const import TEMP_CELSIUS
 from homeassistant.const import (CONF_SCAN_INTERVAL)
+from homeassistant.components.light import Light
 
 _LOGGER = logging.getLogger(__name__)
 
+
 def setup_platform(hass, config, add_devices, add_entities, discovery_info=None):
 	""" Wattio Sensor setup platform """
-	_LOGGER.debug("Wattio Sensor component running ...")
+	_LOGGER.debug("Wattio Switch component running ...")
 	config_path = hass.config.path(WATTIO_CONF_FILE)
 	_LOGGER.debug("Wattio config file: %s" %(config_path))
 	config_status = check_config_file(config_path)
@@ -49,22 +51,8 @@ def setup_platform(hass, config, add_devices, add_entities, discovery_info=None)
 			measurement = None
 			icon = None
 			if device["type"] == "pod":
-				icon = "mdi:power-socket-eu"
-				measurement = "Watt"
-				add_devices([WattioSensor(device["name"],device["type"],measurement,icon,apidata,device["ieee"])], True)
-				_LOGGER.debug("Adding device: %s" %(device["name"]))
-			if device["type"] == "bat":
-				measurement = "Watt"
-				add_devices([WattioSensor(device["name"],device["type"],measurement,icon,apidata,device["ieee"],device["channel"])], True)
-				_LOGGER.debug("Adding device: %s" %(device["name"]))
-			if device["type"] == "therm":
-				measurement = TEMP_CELSIUS
-				add_devices([WattioSensor(device["name"],device["type"],measurement,icon,apidata,device["ieee"])], True)
-				_LOGGER.debug("Adding device: %s" %(device["name"]))
-			if device["type"] == "motion":
-				measurement = TEMP_CELSIUS
-				add_devices([WattioSensor(device["name"],device["type"],measurement,icon,apidata,device["ieee"])], True)
-				_LOGGER.debug("Adding device: %s" %(device["name"]))
+				add_devices([WattioSwitch(device["name"],device["type"],measurement,icon,apidata,device["ieee"])], True)
+				_LOGGER.debug("Adding device: %s" %(device["name"]))	
 	else:
 		''' Not Authorized, need to complete OAUTH2 process '''
 		auth_uri = get_auth_uri(hass,config_file.get("client_id"))
@@ -72,22 +60,38 @@ def setup_platform(hass, config, add_devices, add_entities, discovery_info=None)
 		hass.http.register_view(WattioRegisterView(hass,config,add_entities,config_file.get("client_id"),config_file.get("client_secret"),auth_uri))
 		request_oauth_completion(hass,config,add_devices,add_entities,auth_uri,setup_platform)
 
-class WattioSensor(Entity):
+
+
+class WattioSwitch(Light):
 	""" Representation of Sensor """
 	def __init__(self, name, devtype, measurement,icon,apidata,ieee,channel=None):
 		"""Initialize the sensor."""
 		self._name = name
+		self._is_on = None
 		self._state = None
-		self._measurement = measurement
 		self._icon = icon
 		self._apidata = apidata
 		self._ieee = ieee
 		self._devtype = devtype
 		self._channel = None
-		if channel is not None:
-			self._channel = channel-1
+		self._current_consumption = None
 		#self.type = resource_type
-		#self.update()
+		if self._state == None:
+			self.update()
+
+	@property
+	def is_on(self):
+		return self._state
+	
+	def turn_on(self):
+		self._state = 1
+		_LOGGER.error("Encendiendo")
+		return self._apidata.set_switch_status(self._ieee,"on")
+	
+	def turn_off(self):
+		self._state = 0
+		_LOGGER.error("Apagando")		
+		return self._apidata.set_switch_status(self._ieee,"off")
 
 	@property
 	def name(self):
@@ -95,22 +99,12 @@ class WattioSensor(Entity):
 		return self._name
 
 	@property
-	def unit_of_measurement(self):
-		"""Return the unit of measurement."""
-		return self._measurement
-
-	@property
-	def icon(self):
-		"""Return the image of the sensor."""
-		return self._icon
-
-	@property
-	def state(self):
-		"""Return sensor state"""
-		return self._state
+	def available(self):
+		availability = self._apidata.get_device_availability(self._ieee)
+		return availability
 
 	def update(self):
-		"""Update sensor data."""
-		sensorvalue = self._apidata.update_data(self._name,self._devtype,self._ieee,self._channel)
-		_LOGGER.debug("Updating sensor %s: %s" %(self._name,sensorvalue))
-		self._state = sensorvalue
+		switchvalue = self._apidata.update_switch_status(self._name,self._devtype,self._ieee)
+		_LOGGER.error("%s - %s" %(self._name,switchvalue))
+		self._state = switchvalue
+		return self._state